refactor: move 2048 app to apps directory
